+++ pymc/variational/approximations.py
@@ -19,6 +19,7 @@ from aesara import tensor as at
 from aesara.graph.basic import Variable
 from aesara.tensor.var import TensorVariable
 from arviz import InferenceData
+
 import pymc as pm
 
 from pymc.blocking import DictToArrayBijection
@@ -211,7 +212,9 @@ class EmpiricalGroup(Group):
         trace = self._kwargs.get("trace", None)
         if isinstance(trace, InferenceData):
             raise NotImplementedError(
-                "InferenceData is not supported, use return_inferencedata=False"
+                "The `Empirical` approximation does not yet support `InferenceData` inputs."
+                " Pass `pm.sample(return_inferencedata=False)` to get a `MultiTrace` to use with `Empirical`."
+                " Please help us to refactor: https://github.com/pymc-devs/pymc/issues/5884"
             )
         elif trace is not None and not all(
             [self.model.rvs_to_values[var].name in trace.varnames for var in self.group]